Improve build caching for local runs
Now we cache builds for specific shas.
This avoids the cache where eg we test
pmd/7.0.x against master, and the tool
wants to build at least one of those,
as there is a single pmd-bin-xxx-.zip
in the target dir.

So now there's an archive for each SHA.
This can produce garbage, as now there's
one particular zip for every build, not
one per branch. This should be cleaned
up manually by the developer