Add manual integration test cases
These test cases resemble the use cases how the regression tester is used in the main PMD repo.
Test cases 1-4 are about creating a diff report for various changes: single java rule change, single xpath rule change,
a change in pmd-core and a unrelated change (which means, no diff report is needed).
Test case 5 creates a new baseline.

These tests are not run automatically with `rake integration-test`, since they might break more often. The tests use
the latest PMD changes and modify temporarily some rules on top of that.
You can run these tests with `bundle exec ruby -I test test/manual_integration_tests.rb`. When these tests are green,
there is a good chance, that the integration of the regression tester into the ci scripts in PMD works and provides
the expected results.

The changes on top of the main PMD repo are provided as patch files. These can be used to create a new PR to do
a real test of regression tester on github with the ci scripts.